requirements.txt
# MemoryChat 依赖包

# ==================== 核心依赖 ====================
python-dotenv>=1.0.0

# ==================== LLM 提供商 ====================
# OpenAI (可选,如果使用 OpenAI)
openai>=1.0.0

# Anthropic (可选,如果使用 Claude)
anthropic>=0.25.0

# ==================== Embedding ====================
# Sentence Transformers (推荐,本地 embedding)
sentence-transformers>=2.2.0

# ==================== 向量数据库 ====================
# ChromaDB (推荐)
chromadb>=0.4.0

# ==================== 数据处理 ====================
numpy>=1.24.0

# ==================== 可选依赖 ====================
# Web 框架 (如果开发 API)
# fastapi>=0.100.0
# uvicorn>=0.23.0

# 测试
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
